Initial Agent Implementations
Building off of Ben's work, we now support:

Generating a plan to turn(40) if an agent cannot see the ball, and will
stop (eventually) when it sees the ball. It sometimes takes more than
one full rotation to see the ball though, I'm not sure why.

Support for more than one agent. Look at the .mas2j for some commented
out lines to create more than one agent, and how to log agent behaviour
to a local port.

It seems like the first few messages we receive from the Krislet server
are not related to what the Agent sees, specifically, so I've created
a plan to have the agent wait until it deems the environment to be ready
for it to interact with.

I'm having a hard time ignoring the .class files, so I've had to commit
them as well.