add Keeper Agent to team
This adds a Keeper to the team, which will go to its goal, move out a
little bit, and charge and clear any ball that comes nearby. The Keeper
will ONLY kick straight right now. This means that if the ball gets
behind the Keeper, it will more than likely score an own goal, due to
clearing the ball directly into its own net.

I've been trying to get it to look for the out of bounds line behind it
to get an idea of its facing, but I've been having a hard time getting
line data from the server, I suspect there might be a bug in that the
line data doesn't get correctly saved, or communicated to the players,
or something.